fix: astype(int)-->astype(float) in helper

+++ helper.py
@@ -3,5 +3,5 @@ import numpy as np
 
 
 def invert_signal(sig):
-    sig = np.asarray(sig, dtype=int)
+    sig = np.asarray(sig, dtype=float)
     return sig.max() - sig
